+++ HISTORY.rst
@@ -3,6 +3,15 @@
 Changelog
 =========
 
+
+v0.1.0
+-------
+Release Date: 2021-02-21
+
+    * Bug Fixed:
+        * Using `per` in limitation doesn't work.
+        * Default keys(`limitations`, `key`, `exempt`) doesn't set
+
 v0.0.12
 -------
 Release Date: 2021-02-18